backend/requirements.txt
Flask==3.0.0
flask_cors==3.0.10
waitress==2.0.0
mysql-connector-python==8.2.0
